cdk/text_generation/requirements.txt
boto3
botocore
langchain
langchain-aws
langchain-core
langchain-community
langchain-experimental
langchain-postgres
psycopg2-binary
python-dotenv